Moving and consolidating My Account test fixtures
