+++ src/components/MyAccount/AccountSettingsTab.test.tsx
@@ -1,4 +1,4 @@
-import { patron } from "../../models/modelTests/data/MyAccount"
+import { patron } from "../../../__test__/fixtures/myAccountFixtures"
 import AccountSettingsTab from "./AccountSettingsTab"
 import MyAccount from "../../models/MyAccount"
 import { render, screen } from "../../../src/utils/testUtils"
